+++ InfernalRobotics/InfernalRobotics/Control/ServoGroup.cs
@@ -1,0 +1,25 @@
+﻿using InfernalRobotics.Module;
+
+namespace InfernalRobotics.Control
+{
+    internal class ServoGroup : IServoGroup
+    {
+        private readonly MuMechToggle rawServo;
+
+        public ServoGroup(MuMechToggle rawServo)
+        {
+            this.rawServo = rawServo;
+        }
+
+        public string Name
+        {
+            get { return rawServo.groupName; }
+            set { rawServo.groupName = value; }
+        }
+
+        public float ElectricChargeRequired
+        {
+            get { return rawServo.GroupElectricChargeRequired; }
+        }
+    }
+}
